feat: start proving block asap
