+++ Justfile
@@ -31,27 +31,11 @@ trace-block block chain_id:
     cargo prove --trace 
 
 # Recipe to run the rsp CLI on the latest block in a loop at the given interval and submit proving times to ETH proofs.
-run-eth-proofs cluster_id="1" sleep_time="900":
+run-eth-proofs cluster_id="1" block_interval="100":
     #!/usr/bin/env bash
 
-    while true; do
-        RESPONSE=$(curl -s \
-        -X POST \
-        -H "Content-Type: application/json" \
-        --data '{"jsonrpc":"2.0","method":"eth_blockNumber","params":[],"id":1}' \
-        "$RPC_URL")
-        BLOCK_NUMBER=$((16#$(echo $RESPONSE | grep -o '"result":"[^"]*"' | cut -d'"' -f4 | sed 's/0x//')))
-        echo "Latest block number: $BLOCK_NUMBER"
-
-        ROUNDED_BLOCK=$((BLOCK_NUMBER - (BLOCK_NUMBER % 100)))
-        echo "Rounded block number: $ROUNDED_BLOCK"
-
-        echo "Running rsp..."
-        SP1_PROVER=cuda cargo run --bin rsp --release -F cuda -- --block-number $ROUNDED_BLOCK --eth-proofs-cluster-id {{cluster_id}} --rpc-url $RPC_URL --prove
-
-        echo "Sleeping for $(({{sleep_time}} / 60)) minutes..."
-        sleep {{sleep_time}}
-    done
+    echo "Running rsp..."
+    SP1_PROVER=cuda cargo run --bin eth-proofs --release -F cuda -- --eth-proofs-cluster-id {{cluster_id}} --block-interval {{block_interval}}
 
 # Usage:
 # just run-eth-proofs <cluster-id> <sleep-time>
