+++ bin/eth-proofs/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+version = "0.1.0"
+name = "eth-proofs"
+edition = "2021"
+
+[dependencies]
+tokio.workspace = true
+url.workspace = true
+tracing.workspace = true
+serde_json.workspace = true
+serde.workspace = true
+tracing-subscriber = "0.3.18"
+dotenv = "0.15.0"
+clap = { version = "4.5.7", features = ["derive", "env"] }
+eyre = "0.6.12"
+reqwest = { version =  "0.12.9", features = ["json"] }
+base64 = "0.22.1"
+futures = "0.3"
+
+# workspace
+rsp-host-executor.workspace = true
+rsp-client-executor.workspace = true
+rsp-primitives.workspace = true
+
+# sp1
+sp1-sdk = "4.1.0"
+
+# alloy
+alloy-chains.workspace = true
+alloy-provider = { workspace = true, features = ["ws"] }
+alloy-rpc-client.workspace = true
+alloy-transport.workspace = true
+alloy-transport-ws.workspace = true
+
+[build-dependencies]
+sp1-helper = "4.1.0"
