+++ bin/eth-proofs/build.rs
@@ -1,0 +1,5 @@
+use sp1_helper::build_program;
+
+fn main() {
+    build_program("../client");
+}
