+++ bin/host/Cargo.toml
@@ -9,15 +9,12 @@ url.workspace = true
 tracing-subscriber = "0.3.18"
 dotenv = "0.15.0"
 clap = { version = "4.5.7", features = ["derive", "env"] }
-serde_json.workspace = true
 serde.workspace = true
-bincode = "1.3.3"
 csv = "1.1"
 
 # workspace
 rsp-host-executor.workspace = true
 rsp-client-executor.workspace = true
-rsp-rpc-db.workspace = true
 rsp-primitives.workspace = true
 
 # alloy
@@ -25,22 +22,18 @@ alloy-chains.workspace = true
 alloy-primitives.workspace = true
 alloy-consensus.workspace = true
 alloy-provider.workspace = true
-op-alloy-network.workspace = true
-
 
 # reth 
-reth-evm.workspace = true
 reth-primitives.workspace = true
 reth-primitives-traits.workspace = true
 
 # sp1
-sp1-sdk = "4.0.0"
+sp1-sdk = "4.1.0"
+
 eyre = "0.6.12"
-reqwest = "0.12.9"
-base64 = "0.22.1"
 
 [build-dependencies]
-sp1-helper = "4.0.0"
+sp1-helper = "4.1.0"
 
 [features]
 default = []
